Refactor livolo converters
