+++ src/converters/fromZigbee.ts
@@ -2092,221 +2092,6 @@ export const WSZ01_on_off_action: Fz.Converter<65029, undefined, "attributeRepor
         return {action: `${clickMapping[msg.data["1"]]}`};
     },
 };
-export const livolo_switch_state: Fz.Converter<"genOnOff", undefined, ["attributeReport", "readResponse"]> = {
-    cluster: "genOnOff",
-    type: ["attributeReport", "readResponse"],
-    convert: (model, msg, publish, options, meta) => {
-        const status = msg.data.onOff;
-        return {
-            state_left: status & 1 ? "ON" : "OFF",
-            state_right: status & 2 ? "ON" : "OFF",
-        };
-    },
-};
-export const livolo_socket_state: Fz.Converter<"genPowerCfg", undefined, ["raw"]> = {
-    cluster: "genPowerCfg",
-    type: ["raw"],
-    convert: (model, msg, publish, options, meta) => {
-        const stateHeader = Buffer.from([122, 209]);
-        if (msg.data.indexOf(stateHeader) === 0) {
-            const status = msg.data[14];
-            return {state: status & 1 ? "ON" : "OFF"};
-        }
-    },
-};
-export const livolo_new_switch_state: Fz.Converter<"genPowerCfg", undefined, ["raw"]> = {
-    cluster: "genPowerCfg",
-    type: ["raw"],
-    convert: (model, msg, publish, options, meta) => {
-        const stateHeader = Buffer.from([122, 209]);
-        if (msg.data.indexOf(stateHeader) === 0) {
-            const status = msg.data[14];
-            return {state: status & 1 ? "ON" : "OFF"};
-        }
-    },
-};
-export const livolo_new_switch_state_2gang: Fz.Converter<"genPowerCfg", undefined, ["raw"]> = {
-    cluster: "genPowerCfg",
-    type: ["raw"],
-    convert: (model, msg, publish, options, meta) => {
-        const stateHeader = Buffer.from([122, 209]);
-        if (msg.data.indexOf(stateHeader) === 0) {
-            if (msg.data[10] === 7) {
-                const status = msg.data[14];
-                return {
-                    state_left: status & 1 ? "ON" : "OFF",
-                    state_right: status & 2 ? "ON" : "OFF",
-                };
-            }
-        }
-    },
-};
-export const livolo_new_switch_state_4gang: Fz.Converter<"genPowerCfg", undefined, ["raw"]> = {
-    cluster: "genPowerCfg",
-    type: ["raw"],
-    convert: (model, msg, publish, options, meta) => {
-        const stateHeader = Buffer.from([122, 209]);
-        if (msg.data.indexOf(stateHeader) === 0) {
-            if (msg.data[10] === 7) {
-                const status = msg.data[14];
-                return {
-                    state_left: status & 1 ? "ON" : "OFF",
-                    state_right: status & 2 ? "ON" : "OFF",
-                    state_bottom_left: status & 4 ? "ON" : "OFF",
-                    state_bottom_right: status & 8 ? "ON" : "OFF",
-                };
-            }
-            if (msg.data[10] === 13) {
-                const status = msg.data[13];
-                return {
-                    state_left: status & 1 ? "ON" : "OFF",
-                    state_right: status & 2 ? "ON" : "OFF",
-                    state_bottom_left: status & 4 ? "ON" : "OFF",
-                    state_bottom_right: status & 8 ? "ON" : "OFF",
-                };
-            }
-        }
-    },
-};
-export const livolo_curtain_switch_state: Fz.Converter<"genPowerCfg", undefined, ["raw"]> = {
-    cluster: "genPowerCfg",
-    type: ["raw"],
-    convert: (model, msg, publish, options, meta) => {
-        const stateHeader = Buffer.from([122, 209]);
-        if (msg.data.indexOf(stateHeader) === 0) {
-            if (msg.data[10] === 5 || msg.data[10] === 2) {
-                const status = msg.data[14];
-                return {
-                    state_left: status === 1 ? "ON" : "OFF",
-                    state_right: status === 0 ? "ON" : "OFF",
-                };
-            }
-        }
-    },
-};
-export const livolo_dimmer_state: Fz.Converter<"genPowerCfg", undefined, ["raw"]> = {
-    cluster: "genPowerCfg",
-    type: ["raw"],
-    convert: (model, msg, publish, options, meta) => {
-        const stateHeader = Buffer.from([122, 209]);
-        if (msg.data.indexOf(stateHeader) === 0) {
-            if (msg.data[10] === 7) {
-                const status = msg.data[14];
-                return {state: status & 1 ? "ON" : "OFF"};
-            }
-            if (msg.data[10] === 13) {
-                const status = msg.data[13];
-                return {state: status & 1 ? "ON" : "OFF"};
-            }
-            if (msg.data[10] === 5) {
-                // TODO: Unknown dp, assumed value type
-                const value = msg.data[14] * 10;
-                return {
-                    brightness: mapNumberRange(value, 0, 1000, 0, 255),
-                    brightness_percent: mapNumberRange(value, 0, 1000, 0, 100),
-                    level: value,
-                };
-            }
-        }
-    },
-};
-export const livolo_cover_state: Fz.Converter<"genPowerCfg", undefined, ["raw"]> = {
-    cluster: "genPowerCfg",
-    type: ["raw"],
-    convert: (model, msg, publish, options, meta) => {
-        const dp = msg.data[10];
-        const defaults = {motor_direction: "FORWARD", motor_speed: 40};
-        if (msg.data[0] === 0x7a && msg.data[1] === 0xd1) {
-            const reportType = msg.data[12];
-            switch (dp) {
-                case 0x0c:
-                case 0x0f:
-                    if (reportType === 0x04) {
-                        // Position report
-                        const position = 100 - msg.data[13];
-                        const state = position > 0 ? "OPEN" : "CLOSE";
-                        const moving = dp === 0x0f;
-                        return {...defaults, ...meta.state, position, state, moving};
-                    }
-                    if (reportType === 0x12) {
-                        // Speed report
-                        const motorSpeed = msg.data[13];
-                        return {...defaults, ...meta.state, motor_speed: motorSpeed};
-                    }
-                    if (reportType === 0x13) {
-                        // Direction report
-                        const direction = msg.data[13];
-                        if (direction < 0x80) {
-                            return {...defaults, ...meta.state, motor_direction: "FORWARD"};
-                        }
-                        return {...defaults, ...meta.state, motor_direction: "REVERSE"};
-                    }
-                    break;
-                case 0x02:
-                case 0x03:
-                    // Ignore special commands used only when pairing, as these will rather be handled by `onEvent`
-                    return null;
-                case 0x08:
-                    // Ignore general command acknowledgements, as they provide no useful information.
-                    return null;
-                default:
-                    // Unknown dps
-                    logger.debug(`Unhandled DP ${dp} for ${meta.device.manufacturerName}: ${msg.data.toString("hex")}`, NS);
-            }
-        }
-    },
-};
-export const livolo_hygrometer_state: Fz.Converter<"genPowerCfg", undefined, ["raw"]> = {
-    cluster: "genPowerCfg",
-    type: ["raw"],
-    convert: (model, msg, publish, options, meta) => {
-        const dp = msg.data[10];
-        switch (dp) {
-            case 14:
-                return {
-                    temperature: Number(msg.data[13]),
-                };
-            case 12:
-                return {
-                    humidity: Number(msg.data[13]),
-                };
-        }
-    },
-};
-export const livolo_illuminance_state: Fz.Converter<"genPowerCfg", undefined, ["raw"]> = {
-    cluster: "genPowerCfg",
-    type: ["raw"],
-    convert: (model, msg, publish, options, meta) => {
-        const dp = msg.data[12];
-        const noiseLookup: KeyValueAny = {1: "silent", 2: "normal", 3: "lively", 4: "noisy"};
-        switch (dp) {
-            case 13:
-                return {
-                    illuminance: Number(msg.data[13]),
-                };
-            case 14:
-                return {
-                    noise_detected: msg.data[13] > 2,
-                    noise_level: noiseLookup[msg.data[13]],
-                };
-        }
-    },
-};
-export const livolo_pir_state: Fz.Converter<"genPowerCfg", undefined, ["raw"]> = {
-    cluster: "genPowerCfg",
-    type: ["raw"],
-    convert: (model, msg, publish, options, meta) => {
-        const stateHeader = Buffer.from([122, 209]);
-        if (msg.data.indexOf(stateHeader) === 0) {
-            if (msg.data[10] === 7) {
-                const status = msg.data[14];
-                return {
-                    occupancy: !!(status & 1),
-                };
-            }
-        }
-    },
-};
 export const easycode_action: Fz.Converter<"closuresDoorLock", undefined, "raw"> = {
     cluster: "closuresDoorLock",
     type: "raw",
@@ -2333,114 +2118,6 @@ export const easycodetouch_action: Fz.Converter<"closuresDoorLock", undefined, "
             return {action: value};
         }
         logger.warning(`Unknown lock status with source ${msg.data[3]} and event code ${msg.data[4]}`, NS);
-    },
-};
-export const livolo_switch_state_raw: Fz.Converter<"genPowerCfg", undefined, ["raw"]> = {
-    cluster: "genPowerCfg",
-    type: ["raw"],
-    convert: (model, msg, publish, options, meta) => {
-        /*
-            header                ieee address            info data
-            new socket
-            [124,210,21,216,128,  199,147,3,24,0,75,18,0,  19,7,0]       after interview
-            [122,209,             199,147,3,24,0,75,18,0,  7,1,6,1,0,11] off
-            [122,209,             199,147,3,24,0,75,18,0,  7,1,6,1,1,11] on
-
-            new switch
-            [124,210,21,216,128,  228,41,3,24,0,75,18,0,  19,1,0]       after interview
-            [122,209,             228,41,3,24,0,75,18,0,  7,1,0,1,0,11] off
-            [122,209,             228,41,3,24,0,75,18,0,  7,1,0,1,1,11] on
-
-            old switch
-            [124,210,21,216,128,  170, 10,2,24,0,75,18,0,  17,0,1] after interview
-            [124,210,21,216,0,     18, 15,5,24,0,75,18,0,  34,0,0] left: 0, right: 0
-            [124,210,21,216,0,     18, 15,5,24,0,75,18,0,  34,0,1] left: 1, right: 0
-            [124,210,21,216,0,     18, 15,5,24,0,75,18,0,  34,0,2] left: 0, right: 1
-            [124,210,21,216,0,     18, 15,5,24,0,75,18,0,  34,0,3] left: 1, right: 1
-
-            curtain switch
-            [124,210,21,216,128,  110,74,116,33,0,75,18,0,  19,5,0]        after interview
-            [122,209,             110,74,116,33,0,75,18,0,  5,1,5,0,2,11]  left: 0, right: 0  (off)
-            [122,209,             110,74,116,33,0,75,18,0,  5,1,5,0,1,11]  left: 1, right: 0  (left on)
-            [122,209,             110,74,116,33,0,75,18,0,  5,1,5,0,0,11]  left: 0, right: 1  (right on)
-
-            pir sensor
-            [124,210,21,216,128,  225,52,225,34,0,75,18,0,  19,13,0]       after interview
-            [122,209,             245,94,225,34,0,75,18,0,  7,1,7,1,1,11]  occupancy: true
-            [122,209,             245,94,225,34,0,75,18,0,  7,1,7,1,0,11]  occupancy: false
-
-            hygrometer
-            [122,209,             191,22,3,24,0,75,18,0, 14,1,8,21,14,11]  temperature: 21 degrees Celsius
-            [122,209,             191,22,3,24,0,75,18,0, 12,1,9,73,12,11]  humidity: 73%
-
-            illuminance
-            [124,210,21,216,128,  221,0,115,33,0,75,18,0,  19,12,0]          after interview
-            [122,209,             221,0,115,33,0,75,18,0,  12,1,14,4,12,11]  noise: 4 (noisy)
-            [122,209,             221,0,115,33,0,75,18,0,  12,1,14,3,12,11]  noise: 3 (lively)
-            [122,209,             221,0,115,33,0,75,18,0,  12,1,14,2,12,11]  noise: 2 (normal)
-            [122,209,             221,0,115,33,0,75,18,0,  12,1,14,1,12,11]  noise: 1 (silent)
-            [122,209,             221,0,115,33,0,75,18,0,  12,1,13,20,12,11] lux: 20
-            [122,209,             221,0,115,33,0,75,18,0,  2,0,12,199,1,11]  ??
-            */
-        const malformedHeader = Buffer.from([0x7c, 0xd2, 0x15, 0xd8, 0x00]);
-        const infoHeader = Buffer.from([0x7c, 0xd2, 0x15, 0xd8, 0x80]);
-        // status of old devices
-        if (msg.data.indexOf(malformedHeader) === 0) {
-            const status = msg.data[15];
-            return {
-                state_left: status & 1 ? "ON" : "OFF",
-                state_right: status & 2 ? "ON" : "OFF",
-            };
-        }
-        // info about device
-        if (msg.data.indexOf(infoHeader) === 0) {
-            if (msg.data.includes(Buffer.from([19, 7, 0]), 13)) {
-                // new socket, hack
-                meta.device.modelID = "TI0001-socket";
-                meta.device.save();
-            }
-            // No need to detect this switches, will be done by universal procedure
-            /* if (msg.data.includes(Buffer.from([19, 1, 0]), 13)) {
-                    // new switch, hack
-                    meta.device.modelID = 'TI0001-switch';
-                    meta.device.save();
-                }
-                if (msg.data.includes(Buffer.from([19, 2, 0]), 13)) {
-                    // new switch, hack
-                    meta.device.modelID = 'TI0001-switch-2gang';
-                    meta.device.save();
-                }*/
-            if (msg.data.includes(Buffer.from([19, 5, 0]), 13)) {
-                logger.debug("Detected Livolo Curtain Switch", NS);
-                // curtain switch, hack
-                meta.device.modelID = "TI0001-curtain-switch";
-                meta.device.save();
-            }
-            if (msg.data.includes(Buffer.from([19, 20, 0]), 13)) {
-                // new dimmer, hack
-                meta.device.modelID = "TI0001-dimmer";
-                meta.device.save();
-            }
-            if (msg.data.includes(Buffer.from([19, 21, 0]), 13)) {
-                meta.device.modelID = "TI0001-cover";
-                meta.device.save();
-            }
-            if (msg.data.includes(Buffer.from([19, 13, 0]), 13)) {
-                logger.debug("Detected Livolo Pir Sensor", NS);
-                meta.device.modelID = "TI0001-pir";
-                meta.device.save();
-            }
-            if (msg.data.includes(Buffer.from([19, 15, 0]), 13)) {
-                logger.debug("Detected Livolo Digital Hygrometer", NS);
-                meta.device.modelID = "TI0001-hygrometer";
-                meta.device.save();
-            }
-            if (msg.data.includes(Buffer.from([19, 12, 0]), 13)) {
-                logger.debug("Detected Livolo Digital Illuminance and Sound Sensor", NS);
-                meta.device.modelID = "TI0001-illuminance";
-                meta.device.save();
-            }
-        }
     },
 };
 export const ptvo_switch_uart: Fz.Converter<"genMultistateValue", undefined, ["attributeReport", "readResponse"]> = {
