+++ src/converters/toZigbee.ts
@@ -2244,270 +2244,6 @@ export const humidity: Tz.Converter = {
 // #endregion
 
 // #region Non-generic converters
-export const livolo_socket_switch_on_off: Tz.Converter = {
-    key: ["state"],
-    convertSet: async (entity, key, value, meta) => {
-        if (typeof value !== "string") {
-            return;
-        }
-
-        const state = value.toLowerCase();
-        let oldstate = 1;
-        if (state === "on") {
-            oldstate = 108;
-        }
-        let channel = 1.0;
-        const postfix = meta.endpoint_name || "left";
-        await entity.command("genOnOff", "toggle", {}, {transactionSequenceNumber: 0});
-        const payloadOn = {1: {value: Buffer.from([1, 0, 0, 0, 0, 0, 0, 0]), type: 1}};
-        const payloadOff = {1: {value: Buffer.from([0, 0, 0, 0, 0, 0, 0, 0]), type: 1}};
-        const payloadOnRight = {1: {value: Buffer.from([2, 0, 0, 0, 0, 0, 0, 0]), type: 2}};
-        const payloadOffRight = {1: {value: Buffer.from([0, 0, 0, 0, 0, 0, 0, 0]), type: 2}};
-        const payloadOnBottomLeft = {1: {value: Buffer.from([4, 0, 0, 0, 0, 0, 0, 0]), type: 4}};
-        const payloadOffBottomLeft = {1: {value: Buffer.from([0, 0, 0, 0, 0, 0, 0, 0]), type: 4}};
-        const payloadOnBottomRight = {1: {value: Buffer.from([8, 0, 0, 0, 0, 0, 0, 0]), type: 136}};
-        const payloadOffBottomRight = {1: {value: Buffer.from([0, 0, 0, 0, 0, 0, 0, 0]), type: 136}};
-        if (postfix === "left") {
-            await entity.command("genLevelCtrl", "moveToLevelWithOnOff", {level: oldstate, transtime: channel, optionsMask: 0, optionsOverride: 0});
-            await entity.write("genPowerCfg", state === "on" ? payloadOn : payloadOff, {
-                manufacturerCode: 0x1ad2,
-                disableDefaultResponse: true,
-                disableResponse: true,
-                reservedBits: 3,
-                direction: 1,
-                transactionSequenceNumber: 0xe9,
-            });
-            return {state: {state: value.toUpperCase()}};
-        }
-        if (postfix === "right") {
-            channel = 2.0;
-            await entity.command("genLevelCtrl", "moveToLevelWithOnOff", {level: oldstate, transtime: channel, optionsMask: 0, optionsOverride: 0});
-            await entity.write("genPowerCfg", state === "on" ? payloadOnRight : payloadOffRight, {
-                manufacturerCode: 0x1ad2,
-                disableDefaultResponse: true,
-                disableResponse: true,
-                reservedBits: 3,
-                direction: 1,
-                transactionSequenceNumber: 0xe9,
-            });
-            return {state: {state: value.toUpperCase()}};
-        }
-        if (postfix === "bottom_right") {
-            await entity.write("genPowerCfg", state === "on" ? payloadOnBottomRight : payloadOffBottomRight, {
-                manufacturerCode: 0x1ad2,
-                disableDefaultResponse: true,
-                disableResponse: true,
-                reservedBits: 3,
-                direction: 1,
-                transactionSequenceNumber: 0xe9,
-            });
-            return {state: {state: value.toUpperCase()}};
-        }
-        if (postfix === "bottom_left") {
-            await entity.write("genPowerCfg", state === "on" ? payloadOnBottomLeft : payloadOffBottomLeft, {
-                manufacturerCode: 0x1ad2,
-                disableDefaultResponse: true,
-                disableResponse: true,
-                reservedBits: 3,
-                direction: 1,
-                transactionSequenceNumber: 0xe9,
-            });
-            return {state: {state: value.toUpperCase()}};
-        }
-        return {state: {state: value.toUpperCase()}};
-    },
-    convertGet: async (entity, key, meta) => {
-        await entity.command("genOnOff", "toggle", {}, {transactionSequenceNumber: 0});
-    },
-};
-export const livolo_switch_on_off: Tz.Converter = {
-    key: ["state"],
-    convertSet: async (entity, key, value, meta) => {
-        utils.assertString(value, key);
-        const postfix = meta.endpoint_name || "left";
-        const state = value.toLowerCase() === "on" ? 108 : 1;
-        let channel = 1;
-
-        if (postfix === "left") {
-            channel = 1.0;
-        } else if (postfix === "right") {
-            channel = 2.0;
-        } else {
-            return;
-        }
-
-        await entity.command("genLevelCtrl", "moveToLevelWithOnOff", {level: state, transtime: channel, optionsMask: 0, optionsOverride: 0});
-        return {state: {state: value.toUpperCase()}};
-    },
-    convertGet: async (entity, key, meta) => {
-        await entity.command("genOnOff", "toggle", {}, {transactionSequenceNumber: 0});
-    },
-};
-export const livolo_dimmer_level: Tz.Converter = {
-    key: ["brightness", "brightness_percent", "level"],
-    convertSet: async (entity, key, value, meta) => {
-        // upscale to 100
-        value = Number(value);
-        utils.assertNumber(value, key);
-        // biome-ignore lint/suspicious/noImplicitAnyLet: ignored using `--suppress`
-        let newValue;
-        if (key === "level") {
-            if (value >= 0 && value <= 1000) {
-                newValue = utils.mapNumberRange(value, 0, 1000, 0, 100);
-            } else {
-                throw new Error("Dimmer level is out of range 0..1000");
-            }
-        } else if (key === "brightness_percent") {
-            if (value >= 0 && value <= 100) {
-                newValue = Math.round(value);
-            } else {
-                throw new Error("Dimmer brightness_percent is out of range 0..100");
-            }
-        } else {
-            if (value >= 0 && value <= 255) {
-                newValue = utils.mapNumberRange(value, 0, 255, 0, 100);
-            } else {
-                throw new Error("Dimmer brightness is out of range 0..255");
-            }
-        }
-        await entity.command("genOnOff", "toggle", {}, {transactionSequenceNumber: 0});
-        const payload = {769: {value: Buffer.from([newValue, 0, 0, 0, 0, 0, 0, 0]), type: 1}};
-        await entity.write("genPowerCfg", payload, {
-            manufacturerCode: 0x1ad2,
-            disableDefaultResponse: true,
-            disableResponse: true,
-            reservedBits: 3,
-            direction: 1,
-            transactionSequenceNumber: 0xe9,
-            writeUndiv: true,
-        });
-        return {
-            state: {brightness_percent: newValue, brightness: utils.mapNumberRange(newValue, 0, 100, 0, 255), level: newValue * 10},
-        };
-    },
-    convertGet: async (entity, key, meta) => {
-        await entity.command("genOnOff", "toggle", {}, {transactionSequenceNumber: 0});
-    },
-};
-export const livolo_cover_state: Tz.Converter = {
-    key: ["state"],
-    convertSet: async (entity, key, value, meta) => {
-        utils.assertEndpoint(entity);
-        // biome-ignore lint/suspicious/noImplicitAnyLet: ignored using `--suppress`
-        let payload;
-        const options = {
-            frameType: 0,
-            manufacturerCode: 0x1ad2,
-            disableDefaultResponse: true,
-            disableResponse: true,
-            reservedBits: 3,
-            direction: 1,
-            writeUndiv: true,
-            transactionSequenceNumber: 0xe9,
-        };
-        switch (value) {
-            case "OPEN":
-                payload = {attrId: 0x0000, selector: null, elementData: [0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00]};
-                break;
-            case "CLOSE":
-                payload = {attrId: 0x0000, selector: null, elementData: [0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00]};
-                break;
-            case "STOP":
-                payload = {attrId: 0x0000, selector: null, elementData: [0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00]};
-                break;
-            default:
-                throw new Error(`Value '${value}' is not a valid cover position (must be one of 'OPEN' or 'CLOSE')`);
-        }
-        await entity.writeStructured(
-            "genPowerCfg",
-            // @ts-expect-error workaround write custom payload
-            [payload],
-            options,
-        );
-        return {
-            state: {
-                moving: true,
-            },
-        };
-    },
-};
-export const livolo_cover_position: Tz.Converter = {
-    key: ["position"],
-    convertSet: async (entity, key, value, meta) => {
-        utils.assertNumber(value, key);
-        const position = 100 - value;
-        await entity.command("genOnOff", "toggle", {}, {transactionSequenceNumber: 0});
-        const payload = {1025: {value: [position, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00], type: 1}};
-        await entity.write("genPowerCfg", payload, {
-            manufacturerCode: 0x1ad2,
-            disableDefaultResponse: true,
-            disableResponse: true,
-            reservedBits: 3,
-            direction: 1,
-            transactionSequenceNumber: 0xe9,
-            writeUndiv: true,
-        });
-        return {
-            state: {
-                position: value,
-                moving: true,
-            },
-        };
-    },
-};
-export const livolo_cover_options: Tz.Converter = {
-    key: ["options"],
-    convertSet: async (entity, key, value, meta) => {
-        utils.assertObject(value);
-        const options = {
-            frameType: 0,
-            manufacturerCode: 0x1ad2,
-            disableDefaultResponse: true,
-            disableResponse: true,
-            reservedBits: 3,
-            direction: 1,
-            writeUndiv: true,
-            transactionSequenceNumber: 0xe9,
-        };
-
-        if (value.motor_direction != null) {
-            // biome-ignore lint/suspicious/noImplicitAnyLet: ignored using `--suppress`
-            let direction;
-            switch (value.motor_direction) {
-                case "FORWARD":
-                    direction = 0x00;
-                    break;
-                case "REVERSE":
-                    direction = 0x80;
-                    break;
-                default:
-                    throw new Error(`livolo_cover_options: ${value.motor_direction} is not a valid motor direction \
-                     (must be one of 'FORWARD' or 'REVERSE')`);
-            }
-
-            const payload = {4865: {value: [direction, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00]}};
-            await entity.write(
-                "genPowerCfg",
-                // @ts-expect-error workaround write custom payload
-                payload,
-                options,
-            );
-        }
-
-        if (value.motor_speed != null) {
-            if (value.motor_speed < 20 || value.motor_speed > 40) {
-                throw new Error("livolo_cover_options: Motor speed is out of range (20-40)");
-            }
-            const payload = {4609: {value: [value.motor_speed, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00]}};
-            await entity.write(
-                "genPowerCfg",
-                // @ts-expect-error workaround write custom payload
-                payload,
-                options,
-            );
-        }
-    },
-};
 // biome-ignore lint/style/useNamingConvention: ignored using `--suppress`
 export const ZigUP_lock: Tz.Converter = {
     key: ["led"],
